Sync network reset trigger file path with XAPI
In XAPI 24.22.0-1, that we included in our rebase to 24.39.1-1.1, the
trigger file path was changed from /tmp to /var/tmp, xsconsole was not
updated to sync with that in our package, nor is the one from XenServer
at this time.

Backport the sibbling fix for IPv6 truncaction in pool.conf when using
coordinator IPv6.

Signed-off-by: David Morel <david.morel@vates.tech>